+++ source/js/components/newsletter-signup/molecules/input-checkbox-with-label.jsx
